+++ src/GRA.Controllers/AvatarController.cs
@@ -152,10 +152,9 @@ namespace GRA.Controllers
                             });
                         }
                     }
-                    var siteUrl = await _siteService.GetBaseUrl(Request.Scheme,
-                                Request.Host.Value);
+                    var siteUrl = await _siteLookupService.GetSiteLinkAsync(site.Id);
                     var contentPath = _pathResolver.ResolveContentPath(path);
-                    viewModel.AvatarImageUrl = Path.Combine(siteUrl, contentPath)
+                    viewModel.AvatarImageUrl = Path.Combine(siteUrl.ToString(), contentPath)
                         .Replace("\\", "/", StringComparison.OrdinalIgnoreCase);
 
                     var shareUrl = siteUrl + Url.Action(nameof(ShareController.Avatar), "Share")