+++ CHANGELOG.md
@@ -15,6 +15,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 - At a glance report now works with SQLite
 - Trim coupon codes when they come in from a status update spreadsheet
 - Bug in Remaining Vendor Prize report - deleted users with prizes caused error
+- Remove references to deprecated GetBaseUrl call
 
 ## [4.4.1] 2023-08-11